+++ index.js
@@ -1,5 +1,7 @@
 import bookstoraged from './modules/bookstoraged.js';
 import currenttime from './modules/time.js';
+import { views } from './modules/views.js';
+import add_book from './modules/add_books.js';
 
 class Bookshelf {
   constructor() {
@@ -28,6 +30,7 @@ class Bookshelf {
 }
 
 const books = new Bookshelf();
+
 const bookstorage = JSON.parse(localStorage.getItem('books'));
 if (bookstorage != null) {
   books.book = bookstorage;
@@ -35,79 +38,10 @@ if (bookstorage != null) {
 
 setInterval(currenttime, 1000);
 
-const title = document.querySelector('#book');
-const author = document.querySelector('#author');
 const shelfbook = document.querySelector('.book-shelf');
-const addbtn = document.querySelector('.add-btn');
-const msga = document.querySelector('.msga');
-const msgb = document.querySelector('.msgb');
-const msgc = document.querySelector('.msgc');
-const listtitle = document.querySelector('header');
-const userinput = document.querySelector('.user-input');
-const contact = document.querySelector('.contact');
-const navitems = document.querySelectorAll('.ul-nav li');
-const nav1 = document.querySelector('.nav-items1');
-const nav2 = document.querySelector('.nav-items2');
-const nav3 = document.querySelector('.nav-items3');
 
-navitems.forEach((element, index) => {
-  element.addEventListener('click', () => {
-    if (index === 0) {
-      listtitle.classList.remove('active');
-      shelfbook.classList.remove('active');
-      contact.classList.remove('active');
-      userinput.classList.remove('active');
-      nav1.style.color = 'blue';
-      nav2.style.color = 'black';
-      nav3.style.color = 'black';
-    } else if (index === 1) {
-      listtitle.classList.add('active');
-      shelfbook.classList.add('active');
-      contact.classList.remove('active');
-      userinput.classList.add('active');
-      nav1.style.color = 'black';
-      nav2.style.color = 'blue';
-      nav3.style.color = 'black';
-    } else if (index === 2) {
-      listtitle.classList.add('active');
-      shelfbook.classList.add('active');
-      contact.classList.add('active');
-      userinput.classList.remove('active');
-      nav1.style.color = 'black';
-      nav2.style.color = 'black';
-      nav3.style.color = 'blue';
-    }
-  });
-});
+views();
 
 bookstoraged(books, shelfbook);
 
-function hidemsga() {
-  msga.classList.remove('active');
-}
-function hidemsgb() {
-  msgb.classList.remove('active');
-}
-function hidemsgc() {
-  msgc.classList.remove('active');
-}
-
-addbtn.addEventListener('click', () => {
-  if (title.value === '' || author.value === '') {
-    msga.classList.add('active');
-    setTimeout(hidemsga, 2000);
-  } else if (books.book.filter((element) => element.name === title.value).length !== 0) {
-    msgb.classList.add('active'); setTimeout(hidemsgb, 2000);
-  } else {
-    msgc.classList.add('active');
-    setTimeout(hidemsgc, 2000);
-    books.add(title.value, author.value, shelfbook);
-    const removeBtn = document.querySelectorAll('.remove-btn');
-    removeBtn.forEach((element) => {
-      element.addEventListener('click', () => {
-        const parent = element.parentNode;
-        books.remove(parent);
-      });
-    });
-  }
-});
+add_book(books);
